add fallocate test
This one is meant to pound on the page allocator.  Most of the
will-it-scale test require some kind of copy out to userspace
which is pretty slow.  This should allow any access to the
memory being allocated to be kept inside the kernel.

Note that this will fail in fun ways if you point it at a
filesystem that does not support fallocate(), like ramfs for
instance.

Currently, with tmpfs, this seems to bottleneck on the
counters in __vm_enough_memory() when vm.overcommit_memory=1.

Setting vm.overcommit_memory=0 will make things bottleneck on
the zone->lru_lock from the pagevec freeing code.