tlb_flush1: Seperate file private mapping TLB flush
This test case is designed to measure the performance of tlb flush. Each
thread/process maps a separated private file mapping to memory, and
writes to a page before initiating a tlb flush by madvise(MADV_DONTNEED),
subsequent access of this page will result in repopulating the memory
contents of the up-to-date of the underlying file.

Additionally, it is also a tlb flush sensitive workload to verify the
behavior of tlb hardware, the assertion fails immediately once tlb flush
failure. We have double checked it by hacking Linux kernel to skip tlb
flush manually.

Suggested-by: Dave Hansen <dave.hansen@intel.com>
Signed-off-by: Kemi Wang <kemi.wang@intel.com>