Use hwloc for setting task affinity.
Using sched_set_affinity, sched_get_affinity, and
pthread_attr_setaffinity_np are not portable beyond glibc/Linux. hwloc
has APIs for setting the affinity, so we can use those instead. The
plumbing for doing this for a thread in particular is a little annoying,
as it requires a function prior to testcase_trampoline to set the
affinity.