tlb_flush3: Seperate file shared mapping TLB flush
This test case is designed to measure the performance of tlb flush. Each
thread/process maps a separated shared file mapping to memory, and writes
to a page before initiating a tlb flush by madvise(MADV_DONTNEED),
subsequent access of this page will result in repopulating the memory
contents of the up-to-date version of the underlying file.

Based on the tlb_flush1 test by Kemi Wang.

Signed-off-by: Amitay Isaacs <amitay@ozlabs.org>