poll2: update the limitation when the new limitation is bigger than it
In some case the limitation is set to a very small value, it even smaller
than the old one which will cause "EMFILE" error. For example, when run
"./poll2_threads -t 16 -s 295" on a 4 cpus' machine, the limitation should
 bigger than 128*16=2048, but the new limitation is set to (128+10)*4=552.
So update the limitation only when the new value is bigger than it.

Signed-off-by: Zhengjun Xing <zhengjun.xing@linux.intel.com>