dup1: fix tmpfile leak after exit
Same as unlink1, when the child is killed by the parent,
the /tmp/willitscale-XXXXXX is not removed.

By move the unlink() forward just after the file is created,
we can guarrentee the file will be removed after the process
exit. This will have no impact on the test since the fd is
not closed.

Signed-off-by: Dust Li <dust.li@linux.alibaba.com>