open2: fix tmpfile leak after exit
Same as unlink1, when the child is killed by the parent,
the /tmp/willitscale-XXXXXX/willitscale is not removed.

By moving the tmpdir and tmpfiles creating & removing stage into
testcase_prepare() & testcase_cleanup(), which will run in the
master thread/process, that can make sure all the tmpdirs and tmpfiles
will be deleted after the test.

Signed-off-by: Dust Li <dust.li@linux.alibaba.com>