poll2: update rlim.rlim_cur if it's too low
I saw some assertion failures on a test machine where the hard limit
for open files was adequate but the soft limit was too low. The
startup code only adjust the limits if the hard limit is inadequate.
Make it also check, and if necessary increase, the soft limit.

We make this a separate test so we don't accidentally lower the hard
limit by mistake!

Signed-off-by: Daniel Axtens <dja@axtens.net>