Fix "Assertion `ret == 1' failed" in context_switch1.c
When running the context_switch1 in the thread mode test case, it has
"Assertion `ret == 1' failed" error logs print for the system with a
huge number of CPU cores. The pipe fd(struct args a) init as stack
when the thread number increased to some special number, the stack
data will be damaged, change it to init as the heap, then no assertion
happened again.

Suggested-by: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Zhengjun Xing <zhengjun.xing@linux.intel.com>